+++ assets/plugins/managermanager/install/libs/ddTools/src/ObjectTools/hjson/HJSONException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace HJSON;
+
+class HJSONException extends \Exception
+{
+}